requirements.txt
numpy>=1.26.4
huggingface-hub>=0.21.4
diffusers>=0.26.3
transformers>=4.38.2
torch>=2.2.1
accelerate>=0.26.1
fastapi>=0.110.0
python-dotenv>=1.0.1
uvicorn>=0.27.1
pillow>=10.2.0
omegaconf>=2.3.0
sentencepiece>=0.1.99
beautifulsoup4>=4.12.3
